# Optimize Point class performance (~28% speedup)